Endret slik at easterpage er borte etter easter
