+++ frontend/src/components/pages/EasterPage.tsx
@@ -1,26 +1,5 @@
-import Holidays from 'date-holidays';
 import { useEffect, useState } from 'react';
-// import {useState, useEffect} from 'react';
-
-const countdownToEaster = () => {
-  const today = new Date();
-  const hd = new Holidays("NO");
-  const holidays = hd.getHolidays(today.getFullYear());
-  const goodFridayString = holidays.find(holiday => holiday.name === "Langfredag")?.date;
-  let goodFridayDate;
-  if (goodFridayString) {
-    goodFridayDate = new Date(goodFridayString.replace(" ", "T"));
-  } else {
-    console.log("Good Friday date not found");
-  }
-  const easterEve = new Date(goodFridayDate);
-  easterEve.setDate(goodFridayDate.getDate() + 1);
-  console.log(easterEve);
-
-  const diffTime = Math.abs(easterEve.getTime() - today.getTime());
-  return Math.ceil(diffTime / (1000 * 60 * 60 * 24));
-  };
-
+import countdownToEaster from '../utils/DaysUntilEaster';
 
 export const EasterPage = () => {
   //Const for countdown, use usestate to set "today" as the current date
